Fix the button to work in light mode
