--- web/packages/shared/components/CopyButton.tsx
@@ -1,45 +1,0 @@
-import ContentCopyIcon from "@mui/icons-material/ContentCopy";
-import DoneIcon from "@mui/icons-material/Done";
-import {
-    IconButton,
-    Tooltip,
-    type IconButtonProps,
-    type SvgIconProps,
-} from "@mui/material";
-import { t } from "i18next";
-import { useCallback, useState } from "react";
-
-export default function CopyButton({
-    code,
-    color,
-    size,
-}: {
-    code: string;
-    color?: IconButtonProps["color"];
-    size?: SvgIconProps["fontSize"];
-}) {
-    const [copied, setCopied] = useState(false);
-
-    const handleClick = useCallback(() => {
-        navigator.clipboard.writeText(code);
-        setCopied(true);
-        setTimeout(() => setCopied(false), 1000);
-    }, [code]);
-
-    return (
-        <Tooltip
-            arrow
-            open={copied}
-            title={t("copied")}
-            slotProps={{ popper: { sx: { zIndex: 2000 } } }}
-        >
-            <IconButton onClick={handleClick} color={color}>
-                {copied ? (
-                    <DoneIcon fontSize={size ?? "small"} />
-                ) : (
-                    <ContentCopyIcon fontSize={size ?? "small"} />
-                )}
-            </IconButton>
-        </Tooltip>
-    );
-}
